+++ DSCResources/MSFT_SqlServiceAccount/MSFT_SqlServiceAccount.psm1
@@ -24,8 +24,13 @@ $script:localizedData = Get-LocalizedData -ResourceName 'MSFT_SqlServiceAccount'
         ** Not used in this function **
          Credential of the service account that should be used.
 
+    .PARAMETER VersionNumber
+        ** Only used when specifying IntegrationServices **
+        Version number of IntegrationServices.
+
     .EXAMPLE
         Get-TargetResource -ServerName $env:COMPUTERNAME -InstanceName MSSQLSERVER -ServiceType DatabaseEngine -ServiceAccount $account
+        Get-TargetResource -ServerName $env:COMPUTERNAME -InstanceName MSSQLSERVER -ServiceType IntegrationServices -ServiceAccount $account -VersionNumber 130
 #>
 function Get-TargetResource
 {
@@ -48,11 +53,15 @@ function Get-TargetResource
 
         [Parameter(Mandatory = $true)]
         [System.Management.Automation.PSCredential]
-        $ServiceAccount
+        $ServiceAccount,
+
+        [Parameter(Mandatory=$false)]
+        [System.String]
+        $VersionNumber
     )
 
     # Get the SMO Service object instance
-    $serviceObject = Get-ServiceObject -ServerName $ServerName -InstanceName $InstanceName -ServiceType $ServiceType
+    $serviceObject = Get-ServiceObject -ServerName $ServerName -InstanceName $InstanceName -ServiceType $ServiceType -VersionNumber $VersionNumber
 
     # If no service was found, throw an exception
     if (-not $serviceObject)
@@ -98,8 +107,12 @@ function Get-TargetResource
     .PARAMETER Force
         Forces the service account to be updated.
 
+    .PARAMETER VersionNumber
+        Version number of IntegrationServices
+
     .EXAMPLE
         Test-TargetResource -ServerName $env:COMPUTERNAME -InstanceName MSSQLSERVER -ServiceType DatabaseEngine -ServiceAccount $account
+        Test-TargetResource -ServerName $env:COMPUTERNAME -InstanceName MSSQLSERVER -SerticeType IntegrationServices -ServiceAccount $account -VersionNumber 130
 
 #>
 function Test-TargetResource
@@ -131,7 +144,11 @@ function Test-TargetResource
 
         [Parameter()]
         [System.Boolean]
-        $Force
+        $Force,
+
+        [Parameter()]
+        [System.String]
+        $VersionNumber
     )
 
     if ($Force)
@@ -141,7 +158,7 @@ function Test-TargetResource
     }
 
     # Get the current state
-    $currentState = Get-TargetResource -ServerName $ServerName -InstanceName $InstanceName -ServiceType $ServiceType -ServiceAccount $ServiceAccount
+    $currentState = Get-TargetResource -ServerName $ServerName -InstanceName $InstanceName -ServiceType $ServiceType -ServiceAccount $ServiceAccount -VersionNumber $VersionNumber
     New-VerboseMessage -Message ($script:localizedData.CurrentServiceAccount -f $currentState.ServiceAccountName, $ServerName, $InstanceName)
 
     return ($currentState.ServiceAccountName -ieq $ServiceAccount.UserName)
@@ -170,6 +187,9 @@ function Test-TargetResource
 
     .PARAMETER Force
         Forces the service account to be updated.
+
+    .PARAMETER VersionNumber
+        Version number of IntegrationServices
 
     .EXAMPLE
         Set-TargetResource -ServerName $env:COMPUTERNAME -InstanceName MSSQLSERVER -ServiceType DatabaseEngine -ServiceAccount $account
@@ -202,11 +222,15 @@ function Set-TargetResource
 
         [Parameter()]
         [System.Boolean]
-        $Force
+        $Force,
+
+        [Parameter()]
+        [System.String]
+        $VersionNumber
     )
 
     # Get the Service object
-    $serviceObject = Get-ServiceObject -ServerName $ServerName -InstanceName $InstanceName -ServiceType $ServiceType
+    $serviceObject = Get-ServiceObject -ServerName $ServerName -InstanceName $InstanceName -ServiceType $ServiceType -VersionNumber $VersionNumber
 
     # If no service was found, throw an exception
     if (-not $serviceObject)
@@ -247,9 +271,13 @@ function Set-TargetResource
     .PARAMETER ServiceType
         Type of service to be managed. Must be one of the following:
         DatabaseEngine, SQLServerAgent, Search, IntegrationServices, AnalysisServices, ReportingServices, SQLServerBrowser, NotificationServices.
+    
+    .PARAMETER VersionNumber
+        Version number of IntegrationServices.
 
     .EXAMPLE
         Get-ServiceObject -ServerName $env:COMPUTERNAME -InstanceName MSSQLSERVER -ServiceType DatabaseEngine
+        Get-ServiceObject -ServerName $env:COMPUTERNAME -InstanceName MSSQLSERVER -ServiceType IntegrationServices -VersionNumber 130
 #>
 function Get-ServiceObject
 {
@@ -267,8 +295,19 @@ function Get-ServiceObject
         [Parameter(Mandatory = $true)]
         [ValidateSet('DatabaseEngine', 'SQLServerAgent', 'Search', 'IntegrationServices', 'AnalysisServices', 'ReportingServices', 'SQLServerBrowser', 'NotificationServices')]
         [System.String]
-        $ServiceType
+        $ServiceType,
+
+        [Parameter(Mandatory = $false)]
+        [System.String]
+        $VersionNumber
     )
+
+    # Check to see if Integration services was specified, but no version specified
+    if (($ServiceType -eq "IntegrationServices") -and ([String]::IsNullOrEmpty($VersionNumber)))
+    {
+        $errorMessage = $script:localizedData.UnknownServiceType -f $ServiceType
+        New-InvalidArgumentException -Message $errorMessage -ArgumentName 'VersionNumber'        
+    }
 
     # Load the SMO libraries
     Import-SQLPSModule
@@ -282,9 +321,16 @@ function Get-ServiceObject
     # Get the service name for the specified instance and type
     $serviceNameFilter = Get-SqlServiceName -InstanceName $InstanceName -ServiceType $ServiceType
 
+    # Check the service type and append version number if IntegrationServices
+    if ($ServiceType -eq "IntegrationServices")
+    {
+        # Append version number
+        $serviceNameFilter = $serviceNameFilter + $VersionNumber
+    }
+
     # Get the Service object for the specified instance/type
     $serviceObject = $managedComputer.Services | Where-Object -FilterScript {
-        $_.Name -like "$serviceNameFilter*"
+        $_.Name -eq "$serviceNameFilter"
     }
 
     return $serviceObject
@@ -369,6 +415,9 @@ function ConvertTo-ManagedServiceType
     .PARAMETER ServiceType
         Type of service to be named. Must be one of the following:
         DatabaseEngine, SQLServerAgent, Search, IntegrationServices, AnalysisServices, ReportingServices, SQLServerBrowser, NotificationServices.
+
+    .PARAMETER VersionNumber
+        Version number of IntegrationServices.
 
     .EXAMPLE
         Get-SqlServiceName -InstanceName 'MSSQLSERVER' -ServiceType ReportingServices