+++ Tests/Unit/MSFT_SqlServiceAccount.Tests.ps1
@@ -62,6 +62,8 @@ try
         $mockDefaultServiceAccountCredential = (New-Object -TypeName System.Management.Automation.PSCredential $mockDefaultServiceAccountName, (New-Object -TypeName System.Security.SecureString))
         $mockLocalServiceAccountName = "$($mockSqlServer)\SqlService"
         $mockLocalServiceAccountCredential = (New-Object -TypeName System.Management.Automation.PSCredential $mockLocalServiceAccountName, (New-Object -TypeName System.Security.SecureString))
+        $mockManagedServiceAccountName = 'CONTOSO\sqlservice$'
+        $mockManagedServiceAccountCredential = (New-Object -TypeName System.Management.Automation.PSCredential $mockManagedServiceAccountName, (New-Object -TypeName System.Security.SecureString))
 
         # Stores the result of SetServiceAccount calls
         $testServiceAccountUpdated = @{
@@ -164,6 +166,18 @@ try
                         ServiceAccount = ($mockLocalServiceAccountName -replace $mockSqlServer, '.')
                         Type           = 'SqlServer'
                     }
+
+            $managedComputerObject | Add-Member @mockAddMemberParameters_SetServiceAccount
+
+            return $managedComputerObject
+        }
+
+        $mockGetServiceOBject_DefaultInstance_ManagedServiceAccount = {
+            $managedComputerObject = New-Object -TypeName PSObject -Property @{
+                Name           = $mockDefaultInstanceName
+                ServiceAccount = $mockManagedServiceAccountName
+                Type           = 'SqlServer'      
+            }
 
             $managedComputerObject | Add-Member @mockAddMemberParameters_SetServiceAccount
 
@@ -450,7 +464,8 @@ try
                     },
                     @{
                         ServiceType = 'IntegrationServices'
-                        ExpectedServiceName = 'MsDtsServer'
+                        VersionNumber = '130'
+                        ExpectedServiceName = 'MsDtsServer130'
                     },
                     @{
                         ServiceType = 'AnalysisServices'
@@ -483,7 +498,7 @@ try
                     )
 
                     # Get the service name
-                    Get-SqlServiceName -InstanceName $mockDefaultInstanceName -ServiceType $ServiceType | Should -Be $ExpectedServiceName
+                    Get-SqlServiceName -InstanceName $mockDefaultInstanceName -ServiceType $ServiceType -VersionNumber $VersionNumber | Should -Be $ExpectedServiceName
 
                     # Ensure the mock is utilized
                     Assert-MockCalled -CommandName Get-ChildItem -ParameterFilter $mockGetChildItem_ParameterFilter -Scope It -Exactly -Times 1
@@ -633,6 +648,29 @@ try
                     Assert-MockCalled -CommandName New-Object -Scope It -Exactly -Times 1
                 }
             }
+
+            Context 'When getting service IntegrationServices' {
+                Mock @mockNewObjectParameters_DefaultInstance
+                Mock -CommandName Get-SqlServiceName -MockWith {
+                    return 'MsDtsServer'
+                }
+                It 'Should throw an exception when VersionNumber is not specified'{
+                    $getServiceObjectParameters = $defaultGetServiceObjectParameters.Clone()
+                    $getServiceObjectParameters.ServiceType = 'IntegrationServices'
+
+                    $testErrorMessage = $script:localizedData.UnknownServiceType -f 'IntegrationServices'
+
+                    {Get-ServiceObject @getServiceObjectParameters} | Should -Throw $testErrorMessage
+                }
+
+                It 'Should return service when VersionNumber is specified'{
+                    $getServiceObjectParameters = $defaultGetServiceObjectParameters.Clone()
+                    $getServiceObjectParameters.ServiceType = 'IntegrationServices'
+                    $getServiceObjectParameters.VersionNumber - '130'
+
+                    Get-ServiceObject @getServiceObjectParameters | Should -Be 'MsDtsServer130'
+                }
+            }
         }
 
         Describe 'MSFT_SqlServerServiceAccount\Get-TargetResource' -Tag 'Get' {
@@ -749,6 +787,29 @@ try
 
                     # Ensure mocks were properly used
                     Assert-MockCalled -CommandName Get-ServiceObject -Scope It -Exactly -Times 1
+                }
+            }
+
+            Context 'When the service account is a Managed Service Account' {
+                BeforeAll {
+                    Mock -CommandName Get-ServiceObject -MockWith $mockGetServiceOBject_DefaultInstance_ManagedServiceAccount
+                }
+
+                $defaultGetTargetResourceParameters = @{
+                    ServerName     = $mockSqlServer
+                    InstanceName   = $mockDefaultInstanceName
+                    ServiceType    = $mockServiceType
+                    ServiceAccount = $mockManagedServiceAccountCredential                   
+                }
+
+                It 'Should have the Managed Service Account' {
+                    $currentState = Get-TargetResource @defaultGetTargetResourceParameters
+
+                    # Validate the managed service account
+                    $currentState.ServiceAccountName | Should -Be $mockManagedServiceAccountName
+
+                    # Ensure the mocks were properly used
+                    Asser-MockCalled -CommandName Get-ServiceObject -Scope It -Exactly -Times 1
                 }
             }
         }