+++ CHANGELOG.md
@@ -72,9 +72,9 @@
     when a SQL Server installation is not present on the node running the
     unit test ([issue #983](https://github.com/PowerShell/SqlServerDsc/issues/983)).
 - Changes to SqlServiceAccount
-  - Fixed Get-ServiceObject when searching for IntegrationServices service
+  - Fixed Get-ServiceObject when searching for IntegrationServices service, added VersionNumber.
 - Changes to SqlServiceAccount
-  - Added code to allow for using Managed Service Accounts
+  - Added code to allow for using Managed Service Accounts.
 
 ## 12.0.0.0